UnionFindDecoder: Generalize odd logic
This patch generalizes the is_odd logic, it does that by keeping track of the non_typical nodes in each cluster and then computes is_odd when merging.
Remembering it for each cluster is therefore not necessary anymore.
This allows us to generalise the decoder to codes, where the assumption that a cluster is even when the number of atypical qubits is even doesn't hold.